Renaming cache file
