+++ src/core/operations.ts
@@ -2,7 +2,7 @@ import { Wallet } from '@ethersproject/wallet'
 import { KeyManagementSystem, RIFWallet } from '../lib/core'
 import { getKeys, saveKeys } from '../storage/KeyStore'
 import { Wallets } from '../Context'
-import { Cache } from '../storage/cache/Cache'
+import { Cache } from '../storage/cache/MMKVCache'
 
 export { deleteKeys, hasKeys } from '../storage/KeyStore'
 export { hasPin } from '../storage/PinStore'
